Refactor(performance): Split the memory module
memory.py held 774 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the unit enum
and the stats/snapshot records), loading (LazyProxy, LazyLoader and
StreamProcessor), monitor (MemoryMonitor) and optimizer (the
coordinator, its context manager and factory). The package __init__
re-exports the previous names.

MemoryMonitor's module-level logger moves with it, so the alert test
patches performance.memory.monitor.logger rather than the old
package-level name; patching the package attribute would no longer
reach the global the class actually reads.
